Fix null ref warnings

+++ LcBotCsWeb/Data/Services/DatabaseCache.cs
@@ -6,15 +6,18 @@ namespace LcBotCsWeb.Data.Services;
 
 public class DatabaseCache : ICache
 {
-	private readonly Repository<CachedItem> _collection;
+	private readonly Repository<CachedItem>? _collection;
 
 	public DatabaseCache(Database database)
 	{
 		_collection = database.Cache;
 	}
 
-	public async Task<bool> Create(string key, object obj, TimeSpan timeToLive)
+	public async Task<bool> Create(string key, object? obj, TimeSpan timeToLive)
 	{
+		if (_collection == null)
+			return false;
+
 		if (obj == null)
 			return false;
 
@@ -24,12 +27,18 @@ public class DatabaseCache : ICache
 
 	public async Task<bool> Delete(string key)
 	{
+		if (_collection == null)
+			return false;
+
 		var num = await _collection.Delete(item => item.Key == key);
 		return num > 0;
 	}
 
 	public async Task<T?> Get<T>(string key) where T : class
 	{
+		if (_collection == null)
+			return null;
+
 		var result = (await _collection.Find(item => item.Key == key)).FirstOrDefault();
 
 		if (result == null)
@@ -57,11 +66,17 @@ public class DatabaseCache : ICache
 
 	public async Task Clear()
 	{
+		if (_collection == null)
+			return;
+
 		await _collection.Delete(_ => true);
 	}
 
 	public async Task Cleanup()
 	{
+		if (_collection == null)
+			return;
+
 		await _collection.Delete(item => item.Expires < DateTime.UtcNow);
 	}
 }
